Adjust the code base for 64-bit systems. Fixes #7.
This changes the size of IR from 4 bytes to 8 bytes for 64-bit builds. The critical places during IR code generation have been verified to handle pointer and double values correctly, but since there is no type checking going on in this part of the code and thus relies fully on testing.