Switch to a fixed version of the test suite to avoid externally broken tests.
Since DMDScripts implements an old version of the ECMA standard, newer tests are currently mostly irrelevant for now. However, since we fail a lot of tests, it is important to have a known baseline in terms of which tests succeed. For now, this is only counting the number of passed tests as opposed to tracking each test case individually.